# docs/api.yml
openapi: 3.1.0
info:
  title: Football championship API
  description: Football
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
...
  /matchMaker/{seasonYear}:
    post:
      tags:
        - Match
      summary: Create all matches for a specific season including all clubs
      description: |
        Only season with status STARTED can compute matches, in case season status is either NOT_STARTED or FINISHED,
        must return 400 BAD_REQUEST.
        In case provided season matches already generated, then must return 400 BAD_REQUEST response.
        Otherwise, create two matches for each club against other, one 'HOME' match and one 'AWAY' match.
        Look at the instructions inside the context to have more explanation and examples.
        Also, in case the provided `seasonYear` does not exist, returns 404 NOT_FOUND response.
      parameters:
        - $ref: '#/components/parameters/SeasonYearParameter'
      responses:
        200:
          description: List of created matches for the season
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
  /matches/{seasonYear}:
    get:
      tags:
        - Match
      summary: Get all matches for a specific season
      parameters:
        - $ref: '#/components/parameters/SeasonYearParameter'
        - in: query
          name: matchStatus
          schema:
            $ref: '#/components/schemas/MatchStatus'
          description: |
            Return list filtered by provided matchStatus value. 
            For example, if provided matchStatus = FINISHED, then only FINISHED matches will be returned.
          required: false
        - in: query
          name: clubPlayingName
          schema:
            type: string
          description: |
            Return list filtered by provided clubPlayingName value containing.
            For example, if provided clubPlayingName = 'Real', then return all matches that contains 'Real' as clubs playing name, 
            both if the clubs play HOME or AWAY.
          required: false
        - in: query
          name: matchAfter
          schema:
            type: string
            format: date
          description: |
            Return matches after the provided date only.
            In case both `matchAfter` and `matchBeforeOrEquals` parameters are provided, 
            Return list of matches between date intervals.
          required: false
        - in: query
          name: matchBeforeOrEquals
          schema:
            type: string
            format: date
          description: |
            Return matches before or equals the provided date only.
            In case both `matchAfter` and `matchBeforeOrEquals` parameters are provided, 
            Return list of matches between date intervals.
          required: false
      responses:
        200:
          description: List of matches of the specific season
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
        404:
          $ref: '#/components/responses/NotFoundResponse'
...
components:
  parameters:
...
    SeasonYearParameter:
      in: path
      name: seasonYear
      schema:
        type: string
        format: date
      required: true
  responses:
    NotFoundResponse:
      description: Resource not found
  schemas:
...
    PlayerMinimumInfo:
      properties:
        id:
          type: string
        name:
          type: string
        number:
          type: integer
...
    Match:
      type: object
      properties:
        id:
          type: string
        clubPlayingHome:
          $ref: '#/components/schemas/MatchClub'
        clubPlayingAway:
          $ref: '#/components/schemas/MatchClub'
        stadium:
          type: string
        matchDatetime:
          type: string
          format: date-time
        actualStatus:
          $ref: '#/components/schemas/MatchStatus'
...
    ClubScore:
      type: object
      properties:
        score:
          type: integer
          description: |
            For example, if the match actual score is 'Real Madrid 2 - 3 FC Barcelona', 
            That means Real Madrid has scored 2 two goals and plays at home and FC Barcelona has scored 3 goals and plays away,
            The `home.value` would be 2 and the `away.value' 3.
        scorers:
          type: array
          items:
            $ref: '#/components/schemas/Scorer'
    Scorer:
      type: object
      properties:
        player:
          $ref: '#/components/schemas/PlayerMinimumInfo'
        minuteOfGoal:
          type: integer
          description: Must be between 1 to 90 only, a 400 BAD REQUEST must be given if not.
        ownGoal:
          type: boolean
...
    MatchStatus:
      type: string
      enum:
        - NOT_STARTED
        - STARTED
        - FINISHED
    ClubMinimumInfo:
      type: object
      properties:
        id:
          type: string
          description: Club identifier
        name:
          type: string
        acronym:
          type: string
          description: Three letters describing the club (not-unique)
          example: RMA
    MatchClub:
      allOf:
        - $ref: '#/components/schemas/ClubMinimumInfo'
        - $ref: '#/components/schemas/ClubScore'
...